bugfix(#188): added tracking of api token usage also for teh UrlTokenAuthenticationFilter
